+++ code/src/cljs/sixsq/nuvla/ui/authn/events.cljs
@@ -18,7 +18,7 @@
      [(fn [session]
         (dispatch [::set-session session])
         (when session
-          (dispatch [:sixsq.nuvla.ui.main.events/check-bootstrap-message])
+          #_(dispatch [:sixsq.nuvla.ui.main.events/check-bootstrap-message])
           (dispatch [:sixsq.nuvla.ui.main.events/notifications-polling])))]}))