+++ CHANGELOG.md
@@ -4,16 +4,18 @@
 
 ### Changed
 
+  - BOOTSTRAP_MESSAGE - Disable check for bootstarp message for swarm and 
+    creds
   - DASHBOARD PAGE - Clickable cards better visibility
   - DASHBOARD PAGE - Search Input
   - Main components - Search input
   - EDGE PAGE - Full text search
   - Pagination - make pagination visible even if only one page is available
   - EDGE PAGE - Replace dropdown filter by clickable icons
-  - Credential - add vpn modal, default description and name is set 
+  - CREDENTIAL - add vpn modal, default description and name is set 
     for the user. Warning is displayed for the user to request user 
     to save his generated credential 
-  - Credential - add modal select by default first infra if only 
+  - CREDENTIAL - add modal select by default first infra if only 
     one available
   - OCRE PAGE - add pie chart and align schema with api-server
   - CIMI PAGE - make name field visible by default